+++ changelog.md
@@ -2,6 +2,7 @@
 
 ### 1.7.1
 * Add additional filters allowing for hyphens or underscores to be used when defining. The readme.txt incorrectly used then-unsupported names with hyphens in examples so now we support both.
+* Add numbers="before" and numbers="after" in shortcode
 
 ### 1.7.0
 * Add support for taxonomy term listings to the shortcode